SimpleMessenger: Don't grab the lock when sending messages if we don't have to
We'd like it if sending a message didn't require any global locks, but the
submit_message() function conditionally needs it in order to create new
Pipes. So:
1) Add an "already_locked" param to submit_message; if it is false
and we need to set up a new Pipe we grab (and drop) the lock locally.
2) Have the Connection-based interface set this param to false, and
the addr-based interface set it to true.

Signed-off-by: Somnath Roy <somnath.roy@sandisk.com>
Signed-off-by: Greg Farnum <greg@inktank.com>