IndexManager: Introduced a fast path for getting the file descriptor
Introduced a map for holding the HashIndex objects rather than creating
it every time. After getting the object, it will call fast_lookup on that.
It returns the fd and the full filesystem path (if asked) to the caller.
If there is no upgrade is going on this function is invoked by lfn_open()
in the read path.

Signed-off-by: Somnath Roy <somnath.roy@sandisk.com>
Signed-off-by: Greg Farnum <greg@inktank.com>