PG: The read IO path is calling new getattr/read interface
get_object_context() is modified to take an extra default param
whether snapshot is needed or not. In the find_object_context(), if the
request is coming for only head and it is a read request , we don't need
to go for the get_snapset_context() call. This is the path for the read
IO as well. So, in this case, the new getattr interface is called and
the fd/filepath value are stored within ObjectState class. The same value
is passed during the filestore::read call.

Signed-off-by: Somnath Roy <somnath.roy@sandisk.com>

Conflicts:

	src/osd/ReplicatedPG.cc

Signed-off-by: Greg Farnum <greg@inktank.com>